refactor: common으로 위치 이동 & 이름 변경
