+++ common/src/main/java/com/pickcar/mq/enumType/GpsStatus.java
@@ -1,4 +1,4 @@
-package com.pickcar.domain;
+package com.pickcar.mq.enumType;
 
 public enum GpsStatus {
     NORMAL, ABNORMAL, NOT_INSTALLED, ON_ABNORMAL
